(uniquify; but hey wrong)
This was an attempt at fixing this:

$ radio --dry coul
radio: station key 'coul' ambiguous

which comes from the fact that I have both couleur and couleur3 as
keys and both are found, and both resolve to the same value but I
didn't check for that.

Now I get:

$ radio --dry coul
radio: station key 'coul' not found

hahaha. Since I suppress the double. Ach.