(Better config syntax--broken)
Why?.

chrissbx@novo:~$ radio -l
main>
main> :d $mapping
$VAR1 = sub { "DUMMY" };
main>
Not a HASH reference at /home/chris/TEMP/radio/radio line 137.

How the hell does it come to be that $mapping is a subroutine.

And, why did I have to go with eval at all? require was even
nonworking in an earlier stage of evaluation.
require was nonworking because it didn't see the ogg etc subroutines I think.

Now, if I replace (stations  ) with +{  }, this works here.
So, really. what the.

This is perl, v5.10.0